Rename the crate to `vaccel` and create `ffi` mod
Rename the crate to `vaccel` to reflect that we want to expose
Rust-native primitives instead of direct access to the internal ffi
constructs.

Have the build script write the ffi bindings to a new mod of the crate,
so we can monitor it through git as well.

Finally introduce the SavedModel object and differentiate from Frozen
Graphs.

Signed-off-by: Babis Chalios <mail@bchalios.io>